Implement main game window
Implemented widget that holds the main game where the snake moves. Implemented basic snake movement on arrows and W S A D. Implemented reading game options. Implemented calculations of snake speed without it getting out of hand. Implemented a timer that is set based on the snake speed. Implemented repaint method that paints the snake cells. Parameterized most of everything in the main app so that tweaks are easier.